+++ src/main/java/joshie/harvest/shops/HFShops.java
@@ -9,7 +9,6 @@ import joshie.harvest.api.HFApi;
 import joshie.harvest.api.calendar.CalendarDate;
 import joshie.harvest.api.core.ITiered.ToolTier;
 import joshie.harvest.api.crops.Crop;
-import joshie.harvest.api.shops.IShop;
 import joshie.harvest.buildings.HFBuildings;
 import joshie.harvest.cooking.HFCooking;
 import joshie.harvest.core.HFCore;
@@ -57,17 +56,17 @@ import static joshie.harvest.npc.item.ItemNPCTool.NPCTool.BLUE_FEATHER;
 
 @HFLoader
 public class HFShops {
-    public static final IShop BARN = HFApi.shops.newShop(new ResourceLocation(MODID, "barn"), HFNPCs.BARN_OWNER);
-    public static final IShop CAFE = HFApi.shops.newShop(new ResourceLocation(MODID, "cafe"), HFNPCs.CAFE_OWNER);
-    public static final IShop CARPENTER = HFApi.shops.newShop(new ResourceLocation(MODID, "carpenter"), HFNPCs.BUILDER);
-    public static final IShop POULTRY = HFApi.shops.newShop(new ResourceLocation(MODID, "poultry"), HFNPCs.POULTRY);
-    public static final IShop SUPERMARKET = HFApi.shops.newShop(new ResourceLocation(MODID, "general"), HFNPCs.GS_OWNER);
-    public static final IShop MINER = HFApi.shops.newShop(new ResourceLocation(MODID, "miner"), HFNPCs.MINER);
+    public static final Shop BARN = ShopRegistry.INSTANCE.newShop(new ResourceLocation(MODID, "barn"), HFNPCs.BARN_OWNER);
+    public static final Shop CAFE = ShopRegistry.INSTANCE.newShop(new ResourceLocation(MODID, "cafe"), HFNPCs.CAFE_OWNER);
+    public static final Shop CARPENTER = ShopRegistry.INSTANCE.newShop(new ResourceLocation(MODID, "carpenter"), HFNPCs.BUILDER);
+    public static final Shop POULTRY = ShopRegistry.INSTANCE.newShop(new ResourceLocation(MODID, "poultry"), HFNPCs.POULTRY);
+    public static final Shop SUPERMARKET = ShopRegistry.INSTANCE.newShop(new ResourceLocation(MODID, "general"), HFNPCs.GS_OWNER);
+    public static final Shop MINER = ShopRegistry.INSTANCE.newShop(new ResourceLocation(MODID, "miner"), HFNPCs.MINER);
     //Added in 0.6+
-    public static final IShop BAITSHOP = HFApi.shops.newShop(new ResourceLocation(MODID, "baitshop"), HFNPCs.FISHERMAN);
-    public static final IShop BLOODMAGE = HFApi.shops.newShop(new ResourceLocation(MODID, "bloodmage"), HFNPCs.CLOCKMAKER).setSpecialPurchaseRules(new SpecialRulesFriendship(HFNPCs.CLOCKMAKER, 15000));
-    public static final IShop KITCHEN = HFApi.shops.newShop(new ResourceLocation(MODID, "kitchen"), HFNPCs.CAFE_GRANNY).setSpecialPurchaseRules(new SpecialRulesFriendship(HFNPCs.CAFE_GRANNY, 15000));
-    public static final IShop TRADER = HFApi.shops.newShop(new ResourceLocation(MODID, "trader"), HFNPCs.TRADER).setSpecialPurchaseRules(new SpecialRulesFriendship(HFNPCs.TRADER, 15000));
+    public static final Shop BAITSHOP = ShopRegistry.INSTANCE.newShop(new ResourceLocation(MODID, "baitshop"), HFNPCs.FISHERMAN);
+    public static final Shop BLOODMAGE = ShopRegistry.INSTANCE.newShop(new ResourceLocation(MODID, "bloodmage"), HFNPCs.CLOCKMAKER).setSpecialPurchaseRules(new SpecialRulesFriendship(HFNPCs.CLOCKMAKER, 15000));
+    public static final Shop KITCHEN = ShopRegistry.INSTANCE.newShop(new ResourceLocation(MODID, "kitchen"), HFNPCs.CAFE_GRANNY).setSpecialPurchaseRules(new SpecialRulesFriendship(HFNPCs.CAFE_GRANNY, 15000));
+    public static final Shop TRADER = ShopRegistry.INSTANCE.newShop(new ResourceLocation(MODID, "trader"), HFNPCs.TRADER).setSpecialPurchaseRules(new SpecialRulesFriendship(HFNPCs.TRADER, 15000));
 
     public static void postInit() {
         registerBarn();
@@ -87,58 +86,58 @@ public class HFShops {
         BARN.addPurchasable(1000, HFAnimals.TOOLS.getStackFromEnum(MEDICINE));
         BARN.addPurchasable(new PurchasableEntity(EntityHarvestCow.class, 5000, HFAnimals.ANIMAL.getStackFromEnum(COW), true));
         BARN.addPurchasable(new PurchasableEntity(EntityHarvestSheep.class, 4000, HFAnimals.ANIMAL.getStackFromEnum(SHEEP), true));
-        BARN.addPurchasable(3000, HFAnimals.TOOLS.getStackFromEnum(MIRACLE_POTION));
-        BARN.addPurchasable(500, HFAnimals.TROUGH.getStackFromEnum(WOOD));
+        BARN.addPurchasable(3000, HFAnimals.TOOLS.getStackFromEnum(MIRACLE_POTION), 1);
+        BARN.addPurchasable(500, HFAnimals.TROUGH.getStackFromEnum(WOOD), 3);
         BARN.addPurchasable(250, new ItemStack(Items.NAME_TAG));
         BARN.addPurchasable(10, HFAnimals.TREATS.getStackFromEnum(Treat.GENERIC));
         BARN.addPurchasable(30, HFAnimals.TREATS.getStackFromEnum(Treat.COW));
         BARN.addPurchasable(30, HFAnimals.TREATS.getStackFromEnum(Treat.SHEEP));
-        BARN.addPurchasable(1000, HFAnimals.TOOLS.getStackFromEnum(BRUSH));
-        BARN.addPurchasable(2000, HFAnimals.TOOLS.getStackFromEnum(MILKER));
-        BARN.addPurchasable(1800, new ItemStack(Items.SHEARS));
+        BARN.addPurchasable(1000, HFAnimals.TOOLS.getStackFromEnum(BRUSH), 1);
+        BARN.addPurchasable(2000, HFAnimals.TOOLS.getStackFromEnum(MILKER), 1);
+        BARN.addPurchasable(1800, new ItemStack(Items.SHEARS), 1);
 
         BARN.addOpening(MONDAY, 10000, 15000).addOpening(TUESDAY, 10000, 15000).addOpening(WEDNESDAY, 10000, 15000);
         BARN.addOpening(THURSDAY, 10000, 15000).addOpening(FRIDAY, 10000, 15000).addOpening(SATURDAY, 10000, 15000);
     }
 
     private static void registerBloodmage() {
-        BLOODMAGE.addPurchasable(-30, new ItemStack(Items.SLIME_BALL));
-        BLOODMAGE.addPurchasable(-15, new ItemStack(Items.BONE));
-        BLOODMAGE.addPurchasable(-80, new ItemStack(Items.ENDER_PEARL));
-        BLOODMAGE.addPurchasable(-300, new ItemStack(Items.GOLDEN_APPLE));
-        BLOODMAGE.addPurchasable(-5, new ItemStack(Items.ROTTEN_FLESH));
-        BLOODMAGE.addPurchasable(-10, new ItemStack(Items.SPIDER_EYE));
-        BLOODMAGE.addPurchasable(-30, new ItemStack(Items.POISONOUS_POTATO));
-        BLOODMAGE.addPurchasable(-60, new ItemStack(Items.GHAST_TEAR));
-        BLOODMAGE.addPurchasable(-20, new ItemStack(Items.FERMENTED_SPIDER_EYE));
-        BLOODMAGE.addPurchasable(-70, new ItemStack(Items.BLAZE_ROD));
-        BLOODMAGE.addPurchasable(-35, new ItemStack(Items.MAGMA_CREAM));
-        BLOODMAGE.addPurchasable(-30, new ItemStack(Items.SPECKLED_MELON));
-        BLOODMAGE.addPurchasable(-150, new ItemStack(Items.GOLDEN_CARROT));
-        BLOODMAGE.addPurchasable(-100, new ItemStack(Items.RABBIT_FOOT));
-        BLOODMAGE.addPurchasable(-40, new ItemStack(Items.GUNPOWDER));
-        BLOODMAGE.addPurchasable(-5, new ItemStack(Items.REDSTONE));
-        BLOODMAGE.addPurchasable(-10, new ItemStack(Items.GLOWSTONE_DUST));
+        BLOODMAGE.addPurchasable(-100, new ItemStack(Items.SLIME_BALL), 3);
+        BLOODMAGE.addPurchasable(-150, new ItemStack(Items.BONE), 2);
+        BLOODMAGE.addPurchasable(-800, new ItemStack(Items.ENDER_PEARL), 1);
+        BLOODMAGE.addPurchasable(-300, new ItemStack(Items.GOLDEN_APPLE), 1);
+        BLOODMAGE.addPurchasable(-50, new ItemStack(Items.ROTTEN_FLESH), 5);
+        BLOODMAGE.addPurchasable(-100, new ItemStack(Items.SPIDER_EYE), 2);
+        BLOODMAGE.addPurchasable(-30, new ItemStack(Items.POISONOUS_POTATO), 5);
+        BLOODMAGE.addPurchasable(-600, new ItemStack(Items.GHAST_TEAR), 1);
+        BLOODMAGE.addPurchasable(-150, new ItemStack(Items.FERMENTED_SPIDER_EYE), 1);
+        BLOODMAGE.addPurchasable(-700, new ItemStack(Items.BLAZE_ROD), 1);
+        BLOODMAGE.addPurchasable(-350, new ItemStack(Items.MAGMA_CREAM), 1);
+        BLOODMAGE.addPurchasable(-50, new ItemStack(Items.SPECKLED_MELON), 5);
+        BLOODMAGE.addPurchasable(-150, new ItemStack(Items.GOLDEN_CARROT), 1);
+        BLOODMAGE.addPurchasable(-100, new ItemStack(Items.RABBIT_FOOT), 2);
+        BLOODMAGE.addPurchasable(-200, new ItemStack(Items.GUNPOWDER), 1);
+        BLOODMAGE.addPurchasable(-30, new ItemStack(Items.REDSTONE));
+        BLOODMAGE.addPurchasable(-50, new ItemStack(Items.GLOWSTONE_DUST), 5);
         BLOODMAGE.addOpening(WEDNESDAY, 19000, 24000).addOpening(WEDNESDAY, 0, 5000).addOpening(SATURDAY, 18000, 24000).addOpening(SATURDAY, 0, 3500);
     }
 
     private static void registerCafe() {
         CAFE.addPurchasable(0, new ItemStack(Items.POTIONITEM));
-        CAFE.addPurchasable(new PurchasableMeal(200, new ResourceLocation(MODID, "salad")));
-        CAFE.addPurchasable(new PurchasableMeal(100, new ResourceLocation(MODID, "cookies")));
-        CAFE.addPurchasable(new PurchasableMeal(250, new ResourceLocation(MODID, "juice_pineapple")));
-        CAFE.addPurchasable(new PurchasableMeal(250, new ResourceLocation(MODID, "corn_baked")));
-        CAFE.addPurchasable(new PurchasableMeal(300, new ResourceLocation(MODID, "ice_cream")));
+        CAFE.addPurchasable(new PurchasableMeal(200, new ResourceLocation(MODID, "salad")).setStock(10));
+        CAFE.addPurchasable(new PurchasableMeal(100, new ResourceLocation(MODID, "cookies")).setStock(3));
+        CAFE.addPurchasable(new PurchasableMeal(250, new ResourceLocation(MODID, "juice_pineapple")).setStock(5));
+        CAFE.addPurchasable(new PurchasableMeal(250, new ResourceLocation(MODID, "corn_baked")).setStock(5));
+        CAFE.addPurchasable(new PurchasableMeal(300, new ResourceLocation(MODID, "ice_cream")).setStock(10));
 
         //Allow the purchasing of cookware at the weekends
-        CAFE.addPurchasable(new PurchasableWeekend(25, new ItemStack(COOKBOOK)));
-        CAFE.addPurchasable(new PurchasableWeekend(50, UTENSILS.getStackFromEnum(KNIFE)));
-        CAFE.addPurchasable(new PurchasableWeekend(250, COOKWARE.getStackFromEnum(COUNTER)));
-        CAFE.addPurchasable(new PurchasableWeekend(3000, COOKWARE.getStackFromEnum(FRIDGE)));
-        CAFE.addPurchasable(new PurchasableWeekend(2500, COOKWARE.getStackFromEnum(OVEN_OFF)));
-        CAFE.addPurchasable(new PurchasableWeekend(1500, COOKWARE.getStackFromEnum(FRYING_PAN), COOKWARE.getStackFromEnum(OVEN_OFF)));
-        CAFE.addPurchasable(new PurchasableWeekend(1000, COOKWARE.getStackFromEnum(POT), COOKWARE.getStackFromEnum(OVEN_OFF)));
-        CAFE.addPurchasable(new PurchasableWeekend(1200, COOKWARE.getStackFromEnum(MIXER), COOKWARE.getStackFromEnum(COUNTER)));
+        CAFE.addPurchasable(new PurchasableWeekend(25, new ItemStack(COOKBOOK)).setStock(1));
+        CAFE.addPurchasable(new PurchasableWeekend(50, UTENSILS.getStackFromEnum(KNIFE)).setStock(1));
+        CAFE.addPurchasable(new PurchasableWeekend(250, COOKWARE.getStackFromEnum(COUNTER)).setStock(5));
+        CAFE.addPurchasable(new PurchasableWeekend(3000, COOKWARE.getStackFromEnum(FRIDGE)).setStock(1));
+        CAFE.addPurchasable(new PurchasableWeekend(2500, COOKWARE.getStackFromEnum(OVEN_OFF)).setStock(1));
+        CAFE.addPurchasable(new PurchasableWeekend(1500, COOKWARE.getStackFromEnum(FRYING_PAN), COOKWARE.getStackFromEnum(OVEN_OFF)).setStock(1));
+        CAFE.addPurchasable(new PurchasableWeekend(1000, COOKWARE.getStackFromEnum(POT), COOKWARE.getStackFromEnum(OVEN_OFF)).setStock(1));
+        CAFE.addPurchasable(new PurchasableWeekend(1200, COOKWARE.getStackFromEnum(MIXER), COOKWARE.getStackFromEnum(COUNTER)).setStock(1));
 
         //Add recipes for purchase
         CAFE.addPurchasable(new PurchasableRecipe(SPRING, MONDAY, new ResourceLocation(MODID, "juice_vegetable")));
@@ -165,26 +164,26 @@ public class HFShops {
     }
 
     private static void registerCafeKitchen() {
-        KITCHEN.addPurchasable(-50, new ItemStack(Blocks.BROWN_MUSHROOM));
-        KITCHEN.addPurchasable(-60, new ItemStack(Items.PORKCHOP));
-        KITCHEN.addPurchasable(-80, new ItemStack(Items.COOKED_PORKCHOP));
-        KITCHEN.addPurchasable(-400, new ItemStack(Items.CAKE));
-        KITCHEN.addPurchasable(-250, new ItemStack(Items.COOKIE));
-        KITCHEN.addPurchasable(-60, new ItemStack(Items.BEEF));
-        KITCHEN.addPurchasable(-80, new ItemStack(Items.COOKED_BEEF));
-        KITCHEN.addPurchasable(-40, new ItemStack(Items.CHICKEN));
-        KITCHEN.addPurchasable(-50, new ItemStack(Items.COOKED_CHICKEN));
-        KITCHEN.addPurchasable(-90, new ItemStack(Items.BAKED_POTATO));
-        KITCHEN.addPurchasable(-300, new ItemStack(Items.PUMPKIN_PIE));
-        KITCHEN.addPurchasable(-40, new ItemStack(Items.RABBIT));
-        KITCHEN.addPurchasable(-50, new ItemStack(Items.COOKED_RABBIT));
-        KITCHEN.addPurchasable(-250, new ItemStack(Items.RABBIT_FOOT));
-        KITCHEN.addPurchasable(-80, new ItemStack(Items.MUSHROOM_STEW));
-        KITCHEN.addPurchasable(-80, new ItemStack(Items.MUTTON));
-        KITCHEN.addPurchasable(-100, new ItemStack(Items.COOKED_MUTTON));
-        KITCHEN.addPurchasable(-500, new ItemStack(Items.BEETROOT_SOUP));
-        KITCHEN.addPurchasable(-10, new ItemStack(Items.SUGAR));
-        KITCHEN.addPurchasable(-25, new ItemStack(Items.CHORUS_FRUIT));
+        KITCHEN.addPurchasable(-50, new ItemStack(Blocks.BROWN_MUSHROOM), 5);
+        KITCHEN.addPurchasable(-60, new ItemStack(Items.PORKCHOP), 3);
+        KITCHEN.addPurchasable(-80, new ItemStack(Items.COOKED_PORKCHOP), 4);
+        KITCHEN.addPurchasable(-400, new ItemStack(Items.CAKE), 4);
+        KITCHEN.addPurchasable(-250, new ItemStack(Items.COOKIE), 5);
+        KITCHEN.addPurchasable(-60, new ItemStack(Items.BEEF), 3);
+        KITCHEN.addPurchasable(-80, new ItemStack(Items.COOKED_BEEF), 4);
+        KITCHEN.addPurchasable(-40, new ItemStack(Items.CHICKEN), 5);
+        KITCHEN.addPurchasable(-50, new ItemStack(Items.COOKED_CHICKEN), 5);
+        KITCHEN.addPurchasable(-90, new ItemStack(Items.BAKED_POTATO), 5);
+        KITCHEN.addPurchasable(-300, new ItemStack(Items.PUMPKIN_PIE), 2);
+        KITCHEN.addPurchasable(-40, new ItemStack(Items.RABBIT), 4);
+        KITCHEN.addPurchasable(-50, new ItemStack(Items.COOKED_RABBIT), 5);
+        KITCHEN.addPurchasable(-250, new ItemStack(Items.RABBIT_FOOT), 2);
+        KITCHEN.addPurchasable(-80, new ItemStack(Items.MUSHROOM_STEW), 8);
+        KITCHEN.addPurchasable(-80, new ItemStack(Items.MUTTON), 8);
+        KITCHEN.addPurchasable(-100, new ItemStack(Items.COOKED_MUTTON), 4);
+        KITCHEN.addPurchasable(-500, new ItemStack(Items.BEETROOT_SOUP), 1);
+        KITCHEN.addPurchasable(-10, new ItemStack(Items.SUGAR), 20);
+        KITCHEN.addPurchasable(-25, new ItemStack(Items.CHORUS_FRUIT), 15);
         KITCHEN.addOpening(FRIDAY, 6000, 9500).addOpening(FRIDAY, 17000, 20000);
         HFNPCs.CAFE_GRANNY.setHasInfo(null, null); //Remove the opening times from the granny, it's a bonus
     }
@@ -210,19 +209,19 @@ public class HFShops {
                 CalendarDate date = HFApi.calendar.getDate(world);
                 return (date.getYear() >= 1 || date.getSeason().ordinal() >= 1) && TownHelper.getClosestTownToEntity(player).hasBuilding(HFBuildings.MINING_HILL);
             }
-        }.addTooltip("sprinkler.old"));
+        }.setStock(10).addTooltip("sprinkler.old"));
 
         CARPENTER.addPurchasable(new PurchasableMaterials(10000L, HFCrops.SPRINKLER.getStackFromEnum(Sprinkler.IRON), Logs.of(4), Silver.of(8)) {
             @Override
             public boolean canList(World world, EntityPlayer player) {
                 return HFApi.quests.hasCompleted(Quests.SPRINKLER, player) && TownHelper.getClosestTownToEntity(player).hasBuilding(HFBuildings.MINING_HILL);
             }
-        }.addTooltip("sprinkler.iron.tooltip"));
+        }.setStock(3).addTooltip("sprinkler.iron.tooltip"));
 
         CARPENTER.addPurchasable(new PurchasableMaterials(10, new ItemStack(Blocks.LOG)));
         CARPENTER.addPurchasable(new PurchasableMaterials(20, new ItemStack(Blocks.STONE)));
         if (!HFAnimals.CAN_SPAWN) {
-            CARPENTER.addPurchasable(new PurchasableMaterials(500, new ItemStack(Items.BED), Logs.of(3)));
+            CARPENTER.addPurchasable(new PurchasableMaterials(500, new ItemStack(Items.BED), Logs.of(3)).setStock(1));
         }
 
         //Selling things to the carpenter
@@ -234,15 +233,15 @@ public class HFShops {
     }
 
     private static void registerMiner() {
-        MINER.addPurchasable(1000, HFMining.LADDER.getStackFromEnum(Ladder.DECORATIVE));
-        MINER.addPurchasable(250, HFMining.MINING_TOOL.getStackFromEnum(MiningTool.ESCAPE_ROPE));
+        MINER.addPurchasable(1000, HFMining.LADDER.getStackFromEnum(Ladder.DECORATIVE), 3);
+        MINER.addPurchasable(250, HFMining.MINING_TOOL.getStackFromEnum(MiningTool.ESCAPE_ROPE), 10);
         //Selling things to the mine
-        MINER.addPurchasable(-5, new ItemStack(Items.COAL, 1, 1));
-        MINER.addPurchasable(-8, new ItemStack(Items.GOLD_NUGGET));
-        MINER.addPurchasable(-10, new ItemStack(Items.COAL));
-        MINER.addPurchasable(-15, new ItemStack(Items.QUARTZ));
-        MINER.addPurchasable(-45, new ItemStack(Items.IRON_INGOT));
-        MINER.addPurchasable(-72, new ItemStack(Items.GOLD_INGOT));
+        MINER.addPurchasable(-30, new ItemStack(Items.COAL, 1, 1));
+        MINER.addPurchasable(-15, new ItemStack(Items.GOLD_NUGGET));
+        MINER.addPurchasable(-50, new ItemStack(Items.COAL), 8);
+        MINER.addPurchasable(-60, new ItemStack(Items.QUARTZ), 8);
+        MINER.addPurchasable(-100, new ItemStack(Items.IRON_INGOT), 5);
+        MINER.addPurchasable(-140, new ItemStack(Items.GOLD_INGOT), 2);
 
         MINER.addOpening(MONDAY, 11000, 16000).addOpening(TUESDAY, 11000, 16000).addOpening(WEDNESDAY, 11000, 16000);
         MINER.addOpening(THURSDAY, 11000, 16000).addOpening(FRIDAY, 11000, 16000).addOpening(SATURDAY, 11000, 16000);
@@ -255,9 +254,9 @@ public class HFShops {
         POULTRY.addPurchasable(250, new ItemStack(Items.NAME_TAG));
         POULTRY.addPurchasable(10, HFAnimals.TREATS.getStackFromEnum(Treat.GENERIC));
         POULTRY.addPurchasable(30, HFAnimals.TREATS.getStackFromEnum(Treat.CHICKEN));
-        POULTRY.addPurchasable(500, HFAnimals.TRAY.getStackFromEnum(FEEDER_EMPTY));
-        POULTRY.addPurchasable(1000, HFAnimals.TRAY.getStackFromEnum(NEST_EMPTY));
-        POULTRY.addPurchasable(7500, HFAnimals.SIZED.getStackFromEnum(INCUBATOR));
+        POULTRY.addPurchasable(500, HFAnimals.TRAY.getStackFromEnum(FEEDER_EMPTY), 3);
+        POULTRY.addPurchasable(1000, HFAnimals.TRAY.getStackFromEnum(NEST_EMPTY), 3);
+        POULTRY.addPurchasable(7500, HFAnimals.SIZED.getStackFromEnum(INCUBATOR), 1);
 
         POULTRY.addOpening(MONDAY, 5000, 11000).addOpening(TUESDAY, 5000, 11000).addOpening(WEDNESDAY, 5000, 11000);
         POULTRY.addOpening(THURSDAY, 5000, 11000).addOpening(FRIDAY, 5000, 11000).addOpening(SATURDAY, 5000, 11000);
@@ -270,11 +269,11 @@ public class HFShops {
             }
         }
 
-        SUPERMARKET.addPurchasable(250, HFTools.HOE.getStack(ToolTier.BASIC));
-        SUPERMARKET.addPurchasable(250, HFTools.SICKLE.getStack(ToolTier.BASIC));
-        SUPERMARKET.addPurchasable(500, HFTools.WATERING_CAN.getStack(ToolTier.BASIC));
-        SUPERMARKET.addPurchasable(1000, HFTools.AXE.getStack(ToolTier.BASIC));
-        SUPERMARKET.addPurchasable(1000, HFTools.HAMMER.getStack(ToolTier.BASIC));
+        SUPERMARKET.addPurchasable(250, HFTools.HOE.getStack(ToolTier.BASIC), 1);
+        SUPERMARKET.addPurchasable(250, HFTools.SICKLE.getStack(ToolTier.BASIC), 1);
+        SUPERMARKET.addPurchasable(500, HFTools.WATERING_CAN.getStack(ToolTier.BASIC), 1);
+        SUPERMARKET.addPurchasable(1000, HFTools.AXE.getStack(ToolTier.BASIC), 1);
+        SUPERMARKET.addPurchasable(1000, HFTools.HAMMER.getStack(ToolTier.BASIC), 1);
 
         SUPERMARKET.addPurchasable(new PurchasableBlueFeather(1000, HFNPCs.TOOLS.getStackFromEnum(BLUE_FEATHER)));
         SUPERMARKET.addPurchasable(RICEBALL.getCost(), HFCooking.INGREDIENTS.getStackFromEnum(RICEBALL));
@@ -293,40 +292,40 @@ public class HFShops {
 
     private static void registerTackleshop() {
         BAITSHOP.addPurchasable(new Purchasable(Junk.BAIT.getCost(), HFFishing.JUNK.getStackFromEnum(Junk.BAIT)).addTooltip("junk.bait"));
-        BAITSHOP.addPurchasable(1000L, HFFishing.FISHING_ROD.getStack(ToolTier.BASIC));
-        BAITSHOP.addPurchasable(new Purchasable(1500L, HFFishing.FISHING_BLOCK.getStackFromEnum(FishingBlock.TRAP)).addTooltip("fishing.block.trap"));
+        BAITSHOP.addPurchasable(1000L, HFFishing.FISHING_ROD.getStack(ToolTier.BASIC), 1);
+        BAITSHOP.addPurchasable(new Purchasable(1500L, HFFishing.FISHING_BLOCK.getStackFromEnum(FishingBlock.TRAP)).setStock(10).addTooltip("fishing.block.trap"));
         BAITSHOP.addPurchasable(new PurchasableMaterials(3000L, HFFishing.FISHING_BLOCK.getStackFromEnum(FishingBlock.HATCHERY), Logs.of(8), Wool.of(1)) {
             @Override
             public boolean canList(World world, EntityPlayer player) {
                 return HFApi.quests.hasCompleted(Quests.HATCHERY, player);
             }
-        }.addTooltip("fishing.block.hatchery"));
+        }.setStock(1).addTooltip("fishing.block.hatchery"));
 
 
         //Selling things to the carpenter
-        BAITSHOP.addPurchasable(-100, new ItemStack(Items.PRISMARINE_SHARD));
-        BAITSHOP.addPurchasable(-150, new ItemStack(Items.PRISMARINE_CRYSTALS));
-        BAITSHOP.addPurchasable(-10, new ItemStack(Items.FISH, 1, 0));
-        BAITSHOP.addPurchasable(-30, new ItemStack(Items.FISH, 1, 1));
+        BAITSHOP.addPurchasable(-100, new ItemStack(Items.PRISMARINE_SHARD), 5);
+        BAITSHOP.addPurchasable(-150, new ItemStack(Items.PRISMARINE_CRYSTALS), 3);
+        BAITSHOP.addPurchasable(-10, new ItemStack(Items.FISH, 1, 0), 20);
+        BAITSHOP.addPurchasable(-30, new ItemStack(Items.FISH, 1, 1), 15);
         BAITSHOP.addPurchasable(-50, new ItemStack(Items.FISH, 1, 2));
-        BAITSHOP.addPurchasable(-100, new ItemStack(Items.FISH, 1, 3));
+        BAITSHOP.addPurchasable(-100, new ItemStack(Items.FISH, 1, 3), 5);
         for (Fish fish: Fish.values()) {
             long sell = (fish.getSellValue(fish.getLengthFromSizeOfFish(MEDIUM_FISH))) - fish.getSellValue(fish.getLengthFromSizeOfFish(MEDIUM_FISH)) % 10;
-            BAITSHOP.addPurchasable(new PurchasableObtained(sell, HFFishing.FISH.getStackFromEnum(fish)));
+            BAITSHOP.addPurchasable(new PurchasableObtained(sell, HFFishing.FISH.getStackFromEnum(fish)).setStock(100));
         }
 
         BAITSHOP.addOpening(TUESDAY, 13000, 19000).addOpening(WEDNESDAY, 13000, 19000).addOpening(THURSDAY, 13000, 19000).addOpening(FRIDAY, 13000, 19000);
     }
 
     private static void registerTrader() {
-        TRADER.addPurchasable(-50, new ItemStack(Items.MILK_BUCKET));
-        TRADER.addPurchasable(-10, new ItemStack(Items.EGG));
-        TRADER.addPurchasable(-30, new ItemStack(Items.LEATHER));
-        TRADER.addPurchasable(-20, new ItemStack(Items.FEATHER));
-        TRADER.addPurchasable(-20, new ItemStack(Items.RABBIT_HIDE));
-        TRADER.addPurchasable(-5, new ItemStack(Items.STRING));
+        TRADER.addPurchasable(-60, new ItemStack(Items.MILK_BUCKET), 3);
+        TRADER.addPurchasable(-30, new ItemStack(Items.EGG));
+        TRADER.addPurchasable(-120, new ItemStack(Items.LEATHER), 3);
+        TRADER.addPurchasable(-50, new ItemStack(Items.FEATHER), 8);
+        TRADER.addPurchasable(-80, new ItemStack(Items.RABBIT_HIDE), 5);
+        TRADER.addPurchasable(-20, new ItemStack(Items.STRING));
         for (int i = 0; i < 16; i++) {
-            TRADER.addPurchasable(-15, new ItemStack(Blocks.WOOL, 1, 0));
+            TRADER.addPurchasable(-150, new ItemStack(Blocks.WOOL, 1, 0), 3);
         }
 
         TRADER.addOpening(MONDAY, 6000, 10000).addOpening(TUESDAY, 6000, 10000).addOpening(WEDNESDAY, 6000, 10000).addOpening(THURSDAY, 6000, 10000);