+++ src/main/java/joshie/harvest/shops/ShopRegistry.java
@@ -1,7 +1,6 @@
 package joshie.harvest.shops;
 
 import joshie.harvest.api.npc.INPC;
-import joshie.harvest.api.shops.IShop;
 import joshie.harvest.api.shops.IShopRegistry;
 import joshie.harvest.core.util.annotations.HFApiImplementation;
 import net.minecraft.util.ResourceLocation;
@@ -16,7 +15,7 @@ public class ShopRegistry implements IShopRegistry {
     private ShopRegistry() {}
 
     @Override
-    public IShop newShop(ResourceLocation resource, INPC npc) {
+    public Shop newShop(ResourceLocation resource, INPC npc) {
         Shop shop = new Shop(resource);
         npc.setShop(shop);
         shops.put(resource, shop);